Check ess url before the run

+++ peer/main.go
@@ -79,14 +79,16 @@ func (r *Relay) Init() error {
 	}
 
 	// every hour, delete all very old events
-	go func() {
-		db := r.Storage().(*postgresql.PostgresBackend)
+	if r.ElasticsearchURL == "" {
+		go func() {
+			db := r.Storage().(*postgresql.PostgresBackend)
 
-		for {
-			time.Sleep(60 * time.Minute)
-			db.DB.Exec(`DELETE FROM event WHERE created_at < $1`, time.Now().AddDate(0, -3, 0).Unix()) // 3 months
-		}
-	}()
+			for {
+				time.Sleep(60 * time.Minute)
+				db.DB.Exec(`DELETE FROM event WHERE created_at < $1`, time.Now().AddDate(0, -3, 0).Unix()) // 3 months
+			}
+		}()
+	}
 
 	go func() {
 		ticker := time.NewTicker(3 * time.Second)
